refactor: centralize zone access logic
