+++ src/stores/zoneVisit.ts
@@ -1,28 +1,12 @@
-import type { Zone } from '~/type/zone'
 import { defineStore } from 'pinia'
-import { zonesData } from '~/data/zones'
 import { useShlagedexStore } from './shlagedex'
-import { useZoneProgressStore } from './zoneProgress'
+import { useZoneAccess } from './zoneAccess'
 
 export const useZoneVisitStore = defineStore('zoneVisit', () => {
   const visited = ref<Record<string, boolean>>({})
   const dex = useShlagedexStore()
-  const progress = useZoneProgressStore()
 
-  const zones = zonesData
-  const xpZones = computed(() => zones.filter(z => z.maxLevel > 0))
-
-  function canAccess(z: Zone) {
-    if (z.type === 'village')
-      return z.minLevel <= dex.highestLevel
-    const idx = xpZones.value.findIndex(x => x.id === z.id)
-    if (idx === 0)
-      return true
-    const prev = xpZones.value[idx - 1]
-    return progress.isKingDefeated(prev.id)
-  }
-
-  const accessibleZones = computed(() => zones.filter(z => canAccess(z)))
+  const { accessibleZones } = useZoneAccess(dex.highestLevel)
 
   const hasNewZone = computed(
     () => accessibleZones.value.length > 1
